Helma-AI-Backend/requirements.txt
# Core FastAPI components
fastapi==0.115.12
uvicorn==0.34.1
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
starlette==0.46.1

# Database
motor==3.7.0
pymongo==4.12.0

# Authentication
bcrypt==4.3.0
PyJWT==2.10.1
python-jose==3.4.0
ecdsa==0.19.0
pyasn1==0.4.8
rsa==4.9
six==1.17.0

# Utilities
python-multipart==0.0.20
python-dotenv==1.1.0
email_validator==2.2.0

# Testing
pytest==8.2.0
pytest-asyncio==0.26.0
httpx==0.27.0
httpcore==1.0.8
iniconfig==2.1.0
pluggy==1.5.0
packaging==24.2

# Dependencies
anyio==4.9.0
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
h11==0.14.0
idna==3.10
sniffio==1.3.1
typing_extensions==4.12.2
typing-inspection==0.4.0
annotated-types==0.7.0
certifi==2025.1.31